Create new file list to filter the full file list
The new file added in a patch can be used in the later patches in the
same series. In this case, the file list still contains the new file
introduced in the series and it fails the filter.

This patch creates another list for new files and use it to filter the
full file list.